// src/lib/types/fleet.ts
export interface CreateTripDto {
  vehicleId: string
  destination: string
  purposeOfVisit?: string
  tripDate: string
  meterOut: number
  meterIn?: number
  timeOut?: string
  timeIn?: string
  driverName?: string
  alternativeDriverName?: string
  personTravelList?: string[]
  fuelAllottedLiters?: number
  fuelCostPkr?: number
  status?: string
}